refactor: extract analyze_page_elements runtime
